added support for lmn7.

+++ share/setup.sh
@@ -2,11 +2,13 @@
 # linuxmuster-opsi-setup
 #
 # thomas@linuxmuster.net
-# 20170717
+# 20180216
 #
 
 # read linuxmuster.net environment
 . "$SETTINGS" || exit 1
+# mailip is serverip on lmn62
+grep -q ^mailip "$SETTINGS" || mailip="$serverip"
 
 # test if necessary pkgs are installed
 pkgs_installed &> /dev/null || bailout "The neccessary opsi packages are not installed!"
@@ -131,21 +133,26 @@ create_opsi_cert(){
  local RC="0"
  write_cert_data_to_debconf &> /dev/null || RC="1"
  write_cert_tmp_config || RC="1"
- openssl req -new -x509 -days 1000 -nodes -config "$openssl_conf" -out /etc/opsi/opsiconfd.pem -keyout /etc/opsi/opsiconfd.pem || RC="1"
- openssl gendh -rand $tmp_opsiconfd_rand 512 >> /etc/opsi/opsiconfd.pem || RC="1"
- openssl x509 -subject -dates -fingerprint -noout -in /etc/opsi/opsiconfd.pem || RC="1"
+ openssl req -new -x509 -days 3654 -nodes -config "$openssl_conf" -out "$CONFDPEM" -keyout "$CONFDPEM" || RC="1"
+ openssl gendh -rand $tmp_opsiconfd_rand 512 >> "$CONFDPEM" || RC="1"
+ openssl x509 -subject -dates -fingerprint -noout -in "$CONFDPEM" || RC="1"
  rm -f "$rand_conf" "$openssl_conf"
  return "$RC"
 }
 
 create_certs(){
  local RC="0"
- # ssh
- rm -f /etc/ssh/ssh_host*key*
- ssh-keygen -t dsa -f /etc/ssh/ssh_host_dsa_key -N '' || RC="1"
- ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key -N '' || RC="1"
+ # lmn7
+ if [ -s "$OPSIKEY" -a -s "$OPSICERT" ]; then
+   cat "$OPSIKEY" "$OPSICERT" > "$CONFDPEM" || RC="1"
+ else
+   # lmn62
+   rm -f /etc/ssh/ssh_host*key*
+   ssh-keygen -t dsa -f /etc/ssh/ssh_host_dsa_key -N '' || RC="1"
+   ssh-keygen -t rsa -f /etc/ssh/ssh_host_rsa_key -N '' || RC="1"
+   create_opsi_cert || RC="1"
+ fi
  # opsi
- create_opsi_cert || RC="1"
  return "$RC"
 }
 
@@ -177,7 +184,7 @@ if [ -n "$first" ]; then
          s|@@admin@@|$admin|g
          s|@@domainname@@|$domainname|g" "$PRODCNF_TPL" > "$PRODCNF_TGT" || RC="1"
 else
- sed -e "s|^smtphost .*|smtphost = $serverip|
+ sed -e "s|^smtphost .*|smtphost = $mailip|
          s|^sender .*|sender = opsi-product-updater@$domainname|
          s|^receivers .*|receivers = ${admin}@$domainname|" -i "$PRODCNF_TGT" || RC="1"
 fi
